[MMDB-1443] Now check if given file id is a valid ObjectId in getTags()
Found in testing: given an invalid ObjectId, a runtime exception
("IllegalArgumentException: invalid ObjectId") occurs in
Services.files.getFile().  So check it first.

Made the return values in the tags related code consistent
-- all are now JSON values.  E.g., changed Ok to Ok(toJson("success")).
Changed InternalServerError to NotFound if a file id is not found.

Also made the return values to be consistently JSON values at
3 places in existing code: uploadToDataset(), comment().