[MMDB-1443] added support to get/add/remove/remove_all tags for a dataset
Similar to those for files, the REST end points are:
    GET     /api/datasets/:id/tags               to get tags
    POST    /api/datasets/:id/tags               to add tags
    POST    /api/datasets/:id/tags/remove        to remove tags
    POST    /api/datasets/:id/tags/remove_all    to remove all tags

The existing end point:
    POST    /api/datasets/:id/removeTag
was left for now.  It's slightly diff from the other "remove"
end points: it removes by the ObjectId, which might be needed by
the GUI.  The other "remove" end points remove by value.

The type of "tags" field is List[Tag].  The user id is associated.
In the tests, the user id "anonymous" was automatically set and used.

Tested with curl.  The code also removes the
leading and trailing white spaces (including spaces, tabs, etc.),
and reduces multiple contiguous white spaces to one space character.