[MMDB-1443] tags: data structure change and more error checking
Changed tags from List[String] to List[Tag] for file and section,
changed the html template accordingly.

Added checking of whether a given id is a valid ObjectId, and whether
a file/section/dataset with the given id is found in the DB.  Now
performs the input validation in api/*.scala, and do not check
again in models/*.scala.  This ensures that a correct status is
returned to callers of the REST endpoints, since the models/*.scala
methods usually do not return a status.

Tested with curl.  The userid "anonymous" was set correctly in
the stored tags.