[MMDB-1443] added/improved support for searching by tag
Made changes related to recent data structure changes in file
and section of the "tags" field from List[String] to List[Tag]
in dataset.scala.html and File.scala.

Added the "search by tag" feature for sections, by modifying
Tags.scala: search() and searchByTag.scala.html.

Now the /tags/search/<tag_value> endpoint returns the results
for sections as well, in addition to datasets and files.
The results have a string appended to indicate whether they
match a dataset, a file, or a section, similar to the "Tags"
list in the /datasets/<id> endpoint page.