[MMDB-1444] Added the support of userId and extractor_id for tags
Added extractor_id field in the Tag data structure (models.Tag).
Made it and userId both Option[String], so only non-empty values
are stored in the DB.

Implemented storing userId and extractor_id in tags for files,
datasets, and sections.

Tag's (name, userId, extractor_id) tuple is used as a unique key.
In other words, the same tag names but diff userId or extractor_id
are considered as diff tags.  Current implementation enforces the
restriction which only allows the tags to be removed by the same
user or extractor.

The tags/remove_all REST endpoint is a big hammer -- it does not
check the userId or extractor_id and forcefully remove all tags
for this id.  It is mainly intended for testing.

Changed the tags related code in the dataset.scala.html template
accordingly.