[to-merge-MMDB-1444] Changes due to securesocial API change
securesocial.core.Identity's "id" field is gone in the new version,
and we needed to change to use its other fields.  Use its
"identityId.userId" string for now.  Might change to store other
fields such as the entire identity later, if it's so decided.

This commit is required for the new master branch after Luigi's
recent merge, otherwise it won't compile.