[MMDB-1451] Added support for storing extractor_id for previews, refactored other codes
Added "extractor_id" in models.Preview.

Now we require the extractors to put an "extractor_id" field in the POST data
when posting to attach a preview with a file.  This field is stored in
"previews.files".

Simplified the code: changed "List(file)" to just "file" in controllers.Files.file().