[BD-38] Init REST EP support to get tags and previews for a given file id
Added initial support of a REST endpoint "/api/files/:id/extract"
to get tags and previews for a given file id.  Only a small part
of the whole task.  Both code and route are added.  Tested with
both tags and previews.

Currently the data in the returned JSON object are grouped under
each extractor:
{ "tags": [ {"extractor_id": "e1", "values": ["v1", "v2"]},
            {"extractor_id": "e2", "values": ["v2", "v3"]} ],
  "previews": [ {"extractor_id": "e1",
                 "values": ["preview_id": "id1",
                            "url": "url1",
                            "contentType": "video/mp4"]},
                {"extractor_id": "e2",
                 "values": ["preview_id": "id2",
                            "url": "url2",
                            "contentType": "image/jpg"]} ]
}