[BD-38] Fixed a tag-rel. issue to re-enable tagging area of an image
This feature worked before, but due to the tag related changes
in BD-38 and MMDB-1443, stopped working.  Fixed it and tested to
be working again.

Thanks a lot to Smruti for showing how it works to me, and helping
me locate the code of the feature, so I could try to fix it! :-)