Changes to respond to Rob's comments regarding ensuring that the Rights Holder is displayed fror the varous Creative Commons licenses as well.
In updating to handle this case, some bugs were run into while doing
regression. These included a corner case where reloading the page with the
licenseform on it would cause things to be incorrectly hidden.

Additionally, there was a case where the checkboxes for the Creative
Commons options were not correctly enabling and disabling on data load.

Finally, there is the case where the "I own the rights" box was being
toggled incorrectly at times on reloads of pages, as well as a bug where
the author name for the resource was being referenced in correctly as a
fallback if there is currenly no license data for the rights holder.