Original changes to ensure that the download button is not enabled for users that are not allowed to or when the license specifies that it cannot be downloaded.
Additional changes for MMDB-1598 to ensure that the logic for downloading
a file is present on the server side, where it cannot be bypassed. The
system will also returna  meaningful error message in the case where
download is not allowed.

During the above changes, issue MMDB-1610 was discovered. Updates made to
ensure that if no previewer is available, and if the license does not
allow the raw file to be used, then a reasonable message appears. However,
if the bytes can be downloaded according to the license, the fallback to
usting the file itself as the preview will still happen.