Changes for comment on the pull request to ensure that the email address is used to check in order to display controls for editing and deleting comments.
Additional changes to the comment API service to also check to make sure
that the Identity that is attempting to edit or delete the comment has the
same email as the owner of the comment in order for the request to
proceed.

If the values do not match, the request does not proceed and an error
message is returned.