Changes to incorporate the blueimp opensource file upload library into the system. Upload interface has been changed to utilize the library front end as its basis. Upload controller has been updated to return the appropriate data to feed into the interface.
Modifications made to utilize the appropriate medici APIs and URLs where
necessary. The interface should now support selection of multiple files
via the file browser, drag and drop of one or multiple files onto the
page, and cutting/copying of several files and pasting it into the page.

Once on the page, the interface should allow the user to selectively start
individual files for upload, or start them all at once, or any combination
in between. Additional features are to remove files from the download
queue, cancel an upload that has started, and delete a file from the
system after it has been uploaded (in the case of a mistake for example).
While files are being uploaded, progress bars should be displayed for each
individual file, as well as an overall upload progress bar for the entire
queue.

Additional work was done to enable multiple uploads on a single request. I
rolled those changes back in order to reduce the complexity for the
initial set of changes. Needed to also discuss with others if that is
something that should be supported, and if so, if there are changes to be
made to allow the API side to have a similar feature.

As it stands, most defaults are currently being used for the upload
library. Some of the key ones that can be altered if we feel necessary
are:

type: 'POST'
dataType: 'json'
dropZone: $(document) - jQueryObject, says whole page is drop target
pasteZone: $(document) - jQueryObject, says whole page is paste target
singleFileUploads: true - each file selection is uploaded individually
sequentialUploads: false - If ture, serialized instead of parallel uploads
multipart: true
autoUpload: false - if true, files start uploading as soon as added