Updates to add the ability to delete files from the file detailed view. Addition of new button, placement to prevent accidental deletion, and redirect to appropriate location is all handled.
Additionally, bugs MMDB-1664 and MMDB-1686 are also fixed on this branch.
These items fix the issues with deleting from Tile view and sometimes List
view where the deleted element would still be present until the user
manually refreshed the list. In addition to that, the view that the user
is currently in (tile or list) is the view that is maintained after the
file is deleted from either of those displays.