Updates to create the service calls necessary for a list of datasets that are contained within a space to be retrieved.
Updates to allow the view to correctly list the datasets in the space
detail view. This is a preliminary listing, and is expected to be enhanced
going forward.

Other fixes to allow the cookie for list/tile view mode to propagate to
the Spaces controller, just like it does to all controllers.