Changes to correctly encode the text of tags, as well as to ensure that on adding the tag text immediately after creation, it is not executed.
Other testing to ensure that the delete case is correctly handled, both
immediately after addition, and in the normal case.

It should be noted that there is code that is almost exactly the same
between the two files to handle these items.

Also, the UUID of tags is not used to search and delete currently. It
probably is the safer wa to do it going forward.