Custom enumerator to download large files (1GB+) as a streaming zip file.
This is needed because of output streams not having back pressure. The original output stream was creating data too fast, and play was caching too much data in memory to keep up until it ran out of memory. This enumerator also takes care of stopping the streaming when the browser cancels the download.