list of parent collections and child collections are now UUIDs and not String.
Note - I found some bugs with the space allocation. If a collection was added to a root collection, it was possible to remove it from the root space without removing it from its parent.

Bug found - if you add a collection to a space, the space only increments by one collection unless page is refreshed. in that case the count includes the child collections.