fix to wrongly calculated sha512 for files in mongo store
This only affect the files that are stored in mongo. This code includes
a fix, as well as a wrapper for the code that runs the updates. To run
updates you can use the following code: updateMongo("KEY",
actualFunction). This will make sure that the update is only executed
once iff the MONGOUPDATE flag is set. It will print out the amount of
time an update took.

Also updated logger to add a timestamp for easier debugging.