add active boolean to user
The signup process has changed, all users can now finish the full
signup without any interaction of the admin. Based on the
registerThroughAdmins users are now either immediately activated or are
set to waiting for admins. Every new user registration will result in
an email to  all server admins.

If an account is not active, the user can login, but see nothing (even
if the server is public). Users can see the instance info (from help
menu) and send email to server admins.

Started to remove all use of Identity and switch to our User model.

Removed obsolete code.