added new class vocabularyTerm
The purpose is so that a key can be stored with units (option) or default value.

Example, let's say there is a Vocabulary which represents some kind of manufacturing process. The process will have Terms such as 'material' and a VocabularyTerm such as 'temperature' could contain default units (grad, kelvin) or a default value.