Using single method to create the CONTENT-DISPOSITION header value across all calls that use it.
The updated method should handle spaces in filenames across browsers and
provide an 'ugly but representative' string when special characters
and/or unicode chars are used:

The method first encodes special characters (some of % ,/=&:;?"*) and
then utf-8 encodes the string. Browsers (Chrome, IE, Edge, Firefox) on
Windows appear to decode that differently but ~reasonable way:
simple text with/without spaces appears verbatim
text with special characters mixes decoding verbatim or replacing
disallowed chars with an _ (/\:,?|<> are not allowed on Windows)
adding unicode triggers utf-8 with quoted printable encoding. While
these strings are 'ugly' the quoted printable encoding makes the
ordinary characters visible which helps.

Other notes: 
Clowder itself html escapes characters such as & < and > when you edit a
title and the display and download end up with, for example "&amp;"
instead of &.

Different browsers appear to truncate filenames at different lengths, so
long unicode titles may end up as different filenames on disk depending
on browser for this as well as the decoding differences discussed above.