Updates to support adding an ordered list of creators for datasets and creation objects.
The Dataset and CO models for creators are the same and the display and
edit capabilities are ~consistent across the various pages. 

Note that this update fixes an unreported issue - creator names added to
a Curation Object that contain a comma (e.g. "Bob, Jr.") were split into
multiple creators once saved