return last_average
last_average is only returned if semi binning is enabled, and semi binning could only be spring or summer. Right now if sem is not spring or summer, no data is filtered, but the last_average should be returned.

This should be fixed since it looks like we always want last_average, and so it should not depend on the fact that we set semi to something else than spring/summer.

when fixed this should return last_average: https://seagrant-dev.ncsa.illinois.edu/clowder/api/geostreams/datapoints/trends?sensor_id=950&semi=all