Removed unused RDF libraries. This was probably used by the rdf/xml functionality that was removed a while back but the dependencies were never removed.
Removed Jena validation of JSON-LD metadata. It was creating a blank graph and clients
coulnd't upload metadata when clowder runs in a location that doesn't not have access
to https://clowderframework.org/contexts/metadata.jsonld.

Also removed Jena validation of JSON-LD metadata. It was creating a blank graph and clients
coulnd't upload metadata when clowder runs in a location that doesn't not have access
to https://clowderframework.org/contexts/metadata.jsonld.