initial commit
when tags are added to sections, the new tag cannot be added without the new tag name and id. to fix this, tagService addTagsHelper now also returns the tags that were added in addition to the (not_found, error_string.) So now, adding tags to sections returns a list of the new tags in the response, so that the new tag is added to the right div in sectionRubberband.js.